requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest
backcall==0.1.0           # via ipython
black==19.10b0            # via -r requirements-dev.in
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black
coverage==5.0.4           # via -r requirements-dev.in
decorator==4.4.2          # via ipython, traitlets
dnspython==1.16.0         # via -c requirements.txt, email-validator
email-validator==1.0.5    # via -c requirements.txt, pydantic
entrypoints==0.3          # via flake8
filelock==3.0.12          # via -r requirements-dev.in
flake8==3.7.9             # via -r requirements-dev.in
idna==2.9                 # via -c requirements.txt, email-validator, requests
importlib-metadata==1.6.0  # via pluggy, pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements-dev.in
jedi==0.16.0              # via ipython
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
parso==0.6.2              # via jedi
pathspec==0.7.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.4     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pydantic==1.4             # via -c requirements.txt, -r requirements-dev.in
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython
pyparsing==2.4.6          # via packaging
pytest-django==3.9.0      # via -r requirements-dev.in
pytest==5.4.1             # via pytest-django
regex==2020.2.20          # via black
requests==2.23.0          # via -r requirements-dev.in
six==1.14.0               # via -c requirements.txt, packaging, traitlets
toml==0.10.0              # via black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
urllib3==1.25.8           # via requests
wcwidth==0.1.9            # via prompt-toolkit, pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
